+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Toml config file support to set the auth token
+
 ## [1.4.5] - 2022-01-03
 
 ### Added